resolves #29 rewrite as a proper Asciidoctor converter
- rewrite as a proper Asciidoctor converter
- move main code under the Asciidoctor::Pdf module
- set version to 1.5.0.dev to track core
- update asciidoctor gem to ~> 1.5.0
- migrate to prawn 1.2.1
- add required prawn-table dependency
- make asciidoctor-pdf cli a wrapper around asciidoctor cli
- move inline node handlers to converter
- remove line metrics hacks
  * replace with padding top/bottom adjustment around prose
- evaluate array elements in theme
- update themes
- update examples
- update README
- add Asciidoctor PDF version to PDF header
- take out prose around image logic (moved to editions)
- fix loads of rendering bugs :)
