--- data/templates/inline_button.html.slim
@@ -1,1 +1,0 @@
-b=%([ #{@text} ])
