--- data/templates/inline_footnote.html.slim
@@ -1,5 +1,0 @@
-- if @type == :xref
-  span.footnoteref
-    | [<a class="footnote" href="#_footnote_#{attr :index}" title="View footnote.">#{attr :index}</a>]
-- else
-  | [#{@document.footnotes.find {|fn| fn.index == (attr :index) }.text}]
