--- data/templates/inline_kbd.html.slim
@@ -1,7 +1,0 @@
-- if (keys = attr 'keys').size == 1
-  code=keys.first
-- else
-  - keys.each_with_index do |key, idx|
-    - unless idx.zero?
-      =<> '+'
-    code=key
