--- data/templates/inline_menu.html.slim
@@ -1,15 +1,0 @@
-- menu = attr 'menu'
-- menuitem = attr 'menuitem'
-- if !(submenus = attr 'submenus').empty?
-  strong
-    =menu
-    =<> '|'
-    => submenus.map {|submenu| %(#{submenu} | ) }.join.chop
-    =menuitem
-- elsif !menuitem.nil?
-  strong
-    =menu
-    =<> '|'
-    =menuitem
-- else
-  strong=menu
