+++ lib/asciidoctor-pdf.rb
@@ -1,0 +1,3 @@
+require 'asciidoctor'
+require_relative 'asciidoctor-pdf/version'
+require_relative 'asciidoctor-pdf/converter'
