+++ lib/asciidoctor-pdf/asciidoctor_ext.rb
@@ -1,0 +1,1 @@
+require_relative 'asciidoctor_ext/section'
