+++ lib/asciidoctor-pdf/core_ext/array.rb
@@ -1,5 +1,5 @@
 class Array
   def to_h
-    Hash[*self]
+    Hash[to_a]
   end unless respond_to? :to_h
 end
