+++ lib/asciidoctor-pdf/prawn_ext.rb
@@ -1,0 +1,3 @@
+# the following modules / classes are organized under the Asciidoctor::Prawn namespace
+require_relative 'prawn_ext/extensions'
+require_relative 'prawn_ext/formatted_text/formatter'
