+++ lib/asciidoctor-pdf/prawn_ext/formatted_text/formatter.rb
