+++ lib/asciidoctor-pdf/version.rb
@@ -1,5 +1,5 @@
 module Asciidoctor
 module Pdf
-  VERSION = '1.0.0.dev'
+  VERSION = '1.5.0.dev'
 end
 end
