--- lib/asciidoctor/prawn.rb
@@ -1,2 +1,0 @@
-require 'asciidoctor/prawn/extensions'
-require 'asciidoctor/prawn/formatted_text/formatter'
